Fix tests
And rename the test scripts.
