+++ test/aws-configure.test.js
