+++ test/bnf.test.js
