+++ test/dynamodb-data-type.test.js
