+++ test/dynamodb-delete-item-statement.test.js
